No hard-coded offset for KERNEL_ELF_PADDR_BASE
Commit 2fbbd3c160b46b51d1413c9f91da5bcdf0f7e6bd included this offset to
make room for elfloader, but Sparrow is no longer using that.

sel4/master at time of writing still hard codes this to 0x84000000,
which is still a problem for Sparrow, where that does not refer to
physical memory. (Compare with spike32.dts memory@0x80000000-0xfd000000,
which is probably the platform that upstream is testing.)

This setting is consistent with upstream arm32. The inconsistency
between platforms upstream may be a remnant of bbl or a requirement
imposed by a particular packaging of elfloader.

Change-Id: I27a81557216d7c00aad7f7c2357caa2853cb8875
GitOrigin-RevId: 125ebfdd798dc585ba76f32a591cd9a7e70f9b8e