Allow Lazy SC Rebind
Before, binding a scheduling context to a TCB was not allowed if the
SC was bound to a notification object. Also, binding a SC to a NF was
not allowed if that scheduling context was already bound to a TCB.

Without these restriction it is much easier to move scheduling contexts
around: In effect having a SC bound on both the TCB and a NF acts as if
the thread is running on a donated SC which will be returned when the
tasks calls Recv/Wait, which is done by maybeReturnSchedContext().

This greatly simplifies startup and time-out exception handling for
passive threads having a NF with a SC bound to it.

Signed-off-by: Indan Zupancic <Indan.Zupancic@mep-info.com>

Bug: 250072589

(cherry picked from commit e18e32e28e03e46565f31019da47881a75b764c8)
Change-Id: I9bb1fb47f2458d81c37bc0fdd9db13c3814a7865
GitOrigin-RevId: dacd1d7f1c241542f796e7c75c635207ad9477e5