Do not search for circular includes from depth 0
Switching from gcc 10.1 to 11.1 brings a new entry
(depth=0, header=kernel_out_copy.c). On gcc 10.1, this had been ignored
per the --ignore arg (note the #1 in ignore_re). On gcc 11.1,
kernel_all.i ends up with three lines that start with "# 0":

# 0 "kernel/kernel_all_copy.c"
# 0 "<built-in>"
# 0 "<command-line>"

Probably the interpretation of "depth" is how many includes have been
followed before making it to the filename on that line, so the correct
behavior at depth 0 would be to add the file to the "file_stack".

Change-Id: I9c3fc5e3732c65a78438bb87370ca8e697bea6fe
GitOrigin-RevId: 8e2b95adf2a614e3c2b6f33c79b13f4663ffb2c5