seL4 PLIC lazy interrupt enable
tock is enabling all interrupts for hart 0 and has a default claim.
While we do not want to overwrite settings for IRQs that tock
legitimately needs for itself, we also do not want tock intercepting
IRQs necessary for proper behavior of the software running on the
management core.

This change incorporates a re-revert of
commit 4e0d2fe39b7c7a5b8320463ede6f3934cc6bedec with a fix of the
problem it caused ("tock intercepting" above). It takes the position
that the authority for setting a PLIC register is by default tock but
that for IRQs that are handled on the management core, seL4 is permitted
to and will "steal" those IRQs from all other targets.

Change-Id: I605a83a2dffce0e1a073b248b0f9494ef36c0596
GitOrigin-RevId: 2a76b11c90eac6b69e4fabe5c743971d48b29a39