Update SMC to use mostly standard PLIC.
This CL removes the OT PLIC from Kata in favor of using a RISC-V
compliant PLIC.

However, there is a conflict between HART ids in the standard driver.
From the perspective of the entire system, the SMC is HART 1. However,
from the perspective of the PLIC, the SMC is the only HART and thus HART
0. I've branched riscv_plic0.h to smc_plic and forced the HART id there
to 0.

The removal of lowrisc.h caused bool to be undefined in machine_timer.h,
so I've added a bool typedef there instead.

The device tree has also been modified to match smc.repl.

The machine mode uart now points to uart5 instead.

Change-Id: I25a602e96d3844c7e306aa765d227c9b83040963
GitOrigin-RevId: 9d79615e6c4ca10b681957365a972a4e0898d67e